server/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
pydantic==1.10.14
python-multipart==0.0.9
Pillow==10.3.0
PyYAML==6.0.1
requests==2.31.0
